requirements.txt
# --- Core ETL Needs ---
openpyxl
sqlalchemy
psycopg2-binary
python-dotenv
pandas>=2.2.2
pyarrow>=15.0.2
gspread
oauth2client
# --- Notebook Support ---
jupyterlab       # Or just 'notebook' if you're not using JupyterLab
ipykernel        # Makes the virtualenv usable in Jupyter notebooks

# --- Optional: for dbt integration ---
dbt-core
dbt-postgres


# --- Code Snippets for DevOps dbt, python notebook ---
# dbt run --select property_extract_deduped
